requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
email-validator>=2.1.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2

